ml_microservice/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow-cpu==2.12.0
pydantic==1.10.18
numpy==1.25.2
